updated index.js
